+++ docs/HISTORY.rst
@@ -5,7 +5,8 @@ Changelog
 0.2 (unreleased)
 ~~~~~~~~~~~~~~~~
 
-Nothing yet.
+ - Fixes KML before loading, to avoid common problems found in external
+   KML sources.
 
 0.1 (2014-10-29)
 ~~~~~~~~~~~~~~~~